+++ src/pages/champions/index.ts
@@ -1,0 +1,5 @@
+import mockChampions from "./mockChampions.html";
+
+export {
+    mockChampions,
+};