+++ src/pages/date/index.ts
@@ -1,0 +1,7 @@
+import mockDate20100520 from "./mockDate2010-05-20.html";
+import mockDate20191116 from "./mockDate2019-11-16.html";
+
+export {
+    mockDate20100520,
+    mockDate20191116,
+};