+++ src/pages/events/index.ts
@@ -1,0 +1,11 @@
+import {mockBoutCaneloGGG1} from "./bout";
+import mockEventPageBellewHaye2 from "./mockEventPageBellewHaye2.html";
+import mockEventPageMayweatherMcGregor from "./mockEventPageMayweatherMcGregor.html";
+import mockEventPageNoVenueNoRegionTown from "./mockEventPageNoVenueNoRegionTown.html";
+
+export {
+    mockBoutCaneloGGG1,
+    mockEventPageBellewHaye2,
+    mockEventPageMayweatherMcGregor,
+    mockEventPageNoVenueNoRegionTown,
+};