+++ src/pages/results/index.ts
@@ -1,0 +1,5 @@
+import mockResultsUSMiddleweight from "./mockResultsUSMiddleweight.html";
+
+export {
+    mockResultsUSMiddleweight,
+};