+++ src/pages/schedule/index.ts
@@ -1,0 +1,5 @@
+import mockScheduleWorldwide from "./mockScheduleWorldwide.html";
+
+export {
+    mockScheduleWorldwide,
+};