+++ src/pages/search/index.ts
@@ -1,0 +1,5 @@
+import mockSearchMayweather from "./mockSearchMayweather.html";
+
+export {
+    mockSearchMayweather,
+};