+++ src/pages/title/index.ts
@@ -1,0 +1,5 @@
+import mockMiddleweightWBCBelt from "./mockMiddleweightWBCBelt.html";
+
+export {
+    mockMiddleweightWBCBelt,
+};