+++ src/pages/titles/index.ts
@@ -1,0 +1,9 @@
+import mockTitlesAllTitleAllScheduled from "./mockTitlesAllTitleAllScheduled.html";
+import mockTitlesTitleSelectedAllScheduled from "./mockTitlesTitleSelectedAllScheduled.html";
+import mockTitlesTitleSelectedSuperMiddleweight from "./mockTitlesTitleSelectedSuperMiddleweight.html";
+
+export {
+    mockTitlesAllTitleAllScheduled,
+    mockTitlesTitleSelectedAllScheduled,
+    mockTitlesTitleSelectedSuperMiddleweight
+};