+++ src/pages/venue/index.ts
@@ -1,0 +1,5 @@
+import mockVenueMGMGrand from "./mockVenueMGMGrand.html";
+
+export {
+    mockVenueMGMGrand,
+};